refactor : 자동차 경주 4단계 피드백 반영 리팩토링
