+++ src/main/java/racingcar/domain/policy/MovePolicy.java
@@ -1,0 +1,5 @@
+package racingcar.domain.policy;
+
+public interface MovePolicy {
+    boolean apply();
+}
