--- src/main/java/racingcar/v1/domain/Movable.java
@@ -1,5 +1,0 @@
-package racingcar.v1.domain;
-
-public interface Movable {
-    void forward(boolean value);
-}
