--- src/main/java/racingcar/v1/domain/ParticipantName.java
@@ -1,4 +1,0 @@
-package racingcar.v1.domain;
-
-public class ParticipantName {
-}
