--- src/main/java/racingcar/v1/domain/policy/ZeroToNineRandomPolicy.java
@@ -1,16 +1,0 @@
-package racingcar.v1.domain.policy;
-
-import java.util.Random;
-
-public class ZeroToNineRandomPolicy implements RacingMovePolicy {
-
-    private static final Random random = new Random();
-
-    private static final int FORWARD_CONDITION = 4;
-
-    @Override
-    public boolean execute(int value) {
-        return random.nextInt(value) > FORWARD_CONDITION;
-    }
-
-}
