--- src/main/java/racingcar/v2/domain/policy/MovePolicy.java
@@ -1,5 +1,0 @@
-package racingcar.v2.domain.policy;
-
-public interface MovePolicy {
-    boolean apply(int value);
-}
